Add PsyZ PC port

+++ src/vram.c
@@ -6,7 +6,7 @@ s32 do_load(s16 w, s16 h, s16 x, s16 y, u8 *data)
 
     rect.x = x; rect.y = y;
     rect.w = w; rect.h = h;
-    LoadImage(&rect, (u32 *)data);
+    LoadImage(&rect, (u_long *)data);
     DrawSync(0);
     return w * h * 2;
 }
